IMPALA-8851: Do not throw authorization exception in drop if exists queries
Note that this is the continuation of work in
https://github.com/vihangk1/impala/commits/IMPALA-8851

This patch's goal is to change Impala's behavior in the following case:
- the query is a DROP TABLE/VIEW/DATABASE/FUNCTIONS IF EXISTS statement
- the given object does not exist
- the user has some kind of privilege on the object, which imples the
  privilege to know whether object exists, but does not have DROP
  privilege on the object

Until now this lead to an authorization exception, while it will be
allowed with this change.

An example where this is useful is a user who has CREATE privilege on
a database, and creates table t_owned, and gets ownership of the
table. In this case DROP TABLE IF EXISTS was non idempotent:
DROP TABLE IF EXISTS t_owned;
-> success
DROP TABLE IF EXISTS t_owned;
-> authorization error, as the privileges for the table were
   deleted when the table was successfully dropped

After this change the second statement will be also successful.

The authorization logic has to avoid leaking information that the
user has no right to know. For this reason DROP IF EXISTS has to
return the same error message regardless whether the object exists
or not if the user has no right to know it's existence. This is
achieved with the following pattern:
- in the IF EXISTS case first an ANY privilege is registered, then
  the existence of the object is checked and if it doesn't exist,
  the analysis returns successfully
- if the object exists, the DROP privilege is registered (if there is
  no IF EXISTS in the query, this always happens)
- as the authorization logic checks privileges in the order of
  registration, first the ANY will be checked, and DROP will be only
  checked if the user has ANY privileges

Testing:
- Added a new test case in the sentry tests which confirms that the
authorization exception is not thrown when a drop if exists query is
issued on a object which does not exist.
- Changed several tests affected by the new behavior.
- Ran core tests.

Reviewed-on: http://gerrit.cloudera.org:8080/14121
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 9947e666625474e46f26f0633023beb52577bafa)
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/56011
Tested-by: Jenkins User <jenkins@cloudera.com>
API-Compat: Csaba Ringhofer <csringhofer@cloudera.com>
CDH-Build: Csaba Ringhofer <csringhofer@cloudera.com>
Quasar-L0: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Csaba Ringhofer <csringhofer@cloudera.com>

Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62416
API-Compat: Zsolt Fekete <zfekete@cloudera.com>
CDH-Build: Zsolt Fekete <zfekete@cloudera.com>
Quasar-L0: Zsolt Fekete <zfekete@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Tested-by: Zsolt Fekete <zfekete@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62858

+++ fe/src/main/java/org/apache/impala/analysis/DropTableOrViewStmt.java
@@ -19,7 +19,9 @@ package org.apache.impala.analysis;
 
 import java.util.List;
 
+import org.apache.impala.authorization.AuthorizeableTable;
 import org.apache.impala.authorization.Privilege;
+import org.apache.impala.authorization.PrivilegeRequest;
 import org.apache.impala.catalog.FeTable;
 import org.apache.impala.catalog.FeView;
 import org.apache.impala.catalog.TableLoadingException;
@@ -107,14 +109,25 @@ public class DropTableOrViewStmt extends StatementBase {
     // available in the toThrift() method.
     serverName_ = analyzer.getServerName();
     try {
+      if (ifExists_) {
+        // Start with ANY privilege in case of IF EXISTS, and register DROP privilege
+        // later only if the table exists. See IMPALA-8851 for more explanation.
+        analyzer.registerPrivReq(new PrivilegeRequest(
+            new AuthorizeableTable(dbName_, getTbl()), Privilege.ANY));
+        if (!analyzer.tableExists(tableName_)) return;
+      }
       FeTable table = analyzer.getTable(tableName_, /* add access event */ true,
           /* add column-level privilege */ false, Privilege.DROP);
       Preconditions.checkNotNull(table);
       if (table instanceof FeView && dropTable_) {
+        // DROP VIEW IF EXISTS 'table' succeeds, similarly to Hive, but unlike postgres.
+        if (ifExists_) return;
         throw new AnalysisException(String.format(
             "DROP TABLE not allowed on a view: %s.%s", dbName_, getTbl()));
       }
       if (!(table instanceof FeView) && !dropTable_) {
+        // DROP TABLE IF EXISTS 'view' succeeds, similarly to Hive, but unlike postgres.
+        if (ifExists_) return;
         throw new AnalysisException(String.format(
             "DROP VIEW not allowed on a table: %s.%s", dbName_, getTbl()));
       }
@@ -128,9 +141,6 @@ public class DropTableOrViewStmt extends StatementBase {
           analyzer.getFqTableName(tableName_).toString(), TCatalogObjectType.TABLE,
           Privilege.DROP.toString()));
       LOG.info("Ignoring TableLoadingException for {}", tableName_);
-    } catch (AnalysisException e) {
-      if (!ifExists_) throw e;
-      LOG.info("Ignoring AnalysisException for {}", tableName_);
     }
   }
